requirements.txt
Django>=5.0,<5.1

django-cleanup>=8.1,<9
django-cors-headers>=4.3,<5
django-dramatiq>=0.11,<0.12
django-allauth>=0.63,<1
djangorestframework>=3.15,<4
dj-rest-auth>=6.0,<7

dramatiq[redis]>=1.17,<2
redis>=5,<6

gunicorn>=22,<23
python-dotenv>=1.0,<2
requests>=2.31,<3

numpy>=1.26,<3
torch>=2.3,<3
torchaudio>=2.3,<3
note-seq>=0.0.5
pytube>=15,<16
